+++ editors/code/src/notifications/index.ts
@@ -1,0 +1,5 @@
+import * as publishDecorations from './publish_decorations';
+
+export {
+    publishDecorations,
+};
